Updated model, migration, and validate files

+++ backend/db/migrations/3-create-review.js
@@ -38,7 +38,7 @@ module.exports = {
         allowNull: false
       },
       stars: {
-        type: Sequelize.DECIMAL(1,1),
+        type: Sequelize.INTEGER,
         allowNull: false
       },
       createdAt: {
